requirements.txt
python-telegram-bot==21.6
aiohttp==3.9.5
requests==2.32.3
python-dotenv==1.0.1
